extract getCartCourses out of CartPage
